// src/app/models/Chef.ts
export class Chef {
  id: number | undefined;
  name: string | undefined;
  author: string | undefined;
  description: string | undefined;
  chefPic: string | undefined;
  orginalPrice: number | undefined;
  discountedPrice: number | undefined;
  isInStock: boolean | undefined;
  publishDate: any | undefined;
  // categoryId: number | undefined;
}